refactor: Move react helpers from ListItem into specific file
